+++ Common/BigNumber10Base.cs
@@ -9,7 +9,7 @@ namespace Common
         public BigNumber10Base() : base(new DigitCalculatorGenericBase(10))
         { }
 
-        public BigNumber10Base(uint value) : base(value, new DigitCalculatorGenericBase(10))
+        public BigNumber10Base(ulong value) : base(value, new DigitCalculatorGenericBase(10))
         { }
 
         public BigNumber10Base(List<uint> digits) : base(digits, new DigitCalculatorGenericBase(10))
